refactor: move editing for ast using `SyntaxEditor` to a separate file
